Add debug logs + exception management
Many debug, error and critical log messages have been added, to cope with different situation.
In the query_one_modem function (the one executed in the threads), all code is set in a generic try block in order to catch all exception, thus preventing the process to crash and become a zombie.